tonemapping: more robust detail compression
This revert a part of ef5e67e9bce167c538ccd5291d0f1998d70438b5
The detail layer is not compressed using its min and max anymore:
use its mad instead. To reduce saturation (this was the point of
min and max), use a larger compression factor for the base layer.
Further, the threshold on epsSqrt is set to 150 (for seq flou_IR_X)